refactor: unify credential configuration resolution
Extract credential configuration resolution into a shared
CredentialRequestResolver used by both PresentationDuringIssuanceService
and AuthorizationChallengeEndpoint.

This removes duplicated logic and ensures both components handle
issuer_state, authorization_details, and scope consistently.
